reporter/requirements.txt
fastapi
uvicorn[standard]
confluent-kafka
python-dotenv